Scale subsecond to an 8 bit unsigned integer.

+++ drivers/stm32/rtc.h
@@ -21,7 +21,7 @@ public:
         uint8_t hour;
         uint8_t minute;
         uint8_t second;
-        uint32_t subsecond;
+        uint8_t subsecond;
         AM_PM am_pm;
     };
